\name{graph_data}
\alias{a}
\alias{b}
\docType{data}
\title{
Two randomly shuffled isomorphic graphs
}
\description{
A and B are randomly generated adjacency matrices as shown in the example of run_graph_match function.
a and b are adjacency matrices from the test_simple example in graphm library.
}

\format{

  \describe{
    \item{\code{a}}{an 10x10 adjacency matrix representing graph G to be matched}
    \item{\code{b}}{an  10x10 adjacency matrix representing graph H to be matched}
  }
}

\examples{
print (a)
print (b)


## maybe str(graph_data) ; plot(graph_data) ...
}
\keyword{datasets}
